Remove functions from FileExt
