+++ components/bridge/dao/impl/src/test/java/com/flipperdevices/bridge/dao/impl/md5/MD5FileProviderTest.kt
@@ -2,8 +2,8 @@ package com.flipperdevices.bridge.dao.impl.md5
 
 import android.content.Context
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyContent
-import com.flipperdevices.bridge.dao.impl.FileExt
 import com.flipperdevices.bridge.dao.impl.comparator.DefaultFileComparator
+import com.flipperdevices.bridge.dao.impl.util.TempFolderProvider
 import com.flipperdevices.core.ktx.jre.createNewFileWithMkDirs
 import io.mockk.every
 import io.mockk.mockk
@@ -21,7 +21,7 @@ class MD5FileProviderTest {
     @Before
     fun setUp() {
         context = mockk()
-        tempFolder = FileExt.provideTempFolder()
+        tempFolder = TempFolderProvider.provide()
         every { context.filesDir } returns tempFolder
     }
